Fix: Correct discount badge visibility logic
Corrected the logic for showing the discount badge by casting the `discount` value to an integer before the comparison. This ensures that the badge is hidden for discount values that are greater than 0 but less than 1, which should not be displayed.

+++ lib/widget/product_item.dart
@@ -12,7 +12,7 @@ class ProductItem extends StatelessWidget {
 
   // --- 1. Discount Badge Helper ---
   Widget _buildDiscountBadge(BuildContext context, double discount) {
-    if (discount <= 0) return const SizedBox.shrink();
+    if (discount.toInt() <= 0) return const SizedBox.shrink();
     return Positioned(
       top: 0,
       right: 0,
